requirements.txt
Django==4.1.9
django-phonenumber-field==7.1.0
django-filter==23.2
django-select2==8.1.2
phonenumbers==8.13.13
mssql-django==1.3